model: Changes modelBase.validate to be async
Added a method to ModelValidation that will be used for querying the api/schemas endpoint.
This will be an async operation and therefore the validate() method should return a promise.
ModelBase.validate() might now return a false positive if as there are not yet messages returned
by the ModelValidation.validate method.

References #9
References #11