Client side model validation messages; E2E tests
Fixes the E2E tests. They are still however needed to be included in
the build.
ModelValidation.validate() now returns a status object with the
status and the messages that are relevant when that status is false.

Both the async validate against the api/schemas endpoint and the client
side validation is combined for the end result of the ModelBase.validate()
result.

Closes #10
References #11