[tools] Correct faulty jspm_packages path.
The end to end tests did not run correctly on travis due to the
jspm_packages folder being outside of the root project folder.
Changed the baseURL for jspm to be in the root of the project and
adjusted the karma.conf.js and gulp build files to adjust for this
change.

References #4