[docs] Update docs with new code example runner
Code examples can now also be defined as runnable.
A new runner is included that can run the examples in a
webworker against the defined dhis2 instance. By
default this is defined as the demo instance.

We need to make sure that the remote access setting is
set correctly in that instance for this to work.

References #5