[docs] Add static pages and runnable examples
Implemented support for runnable examples. For an example
to be marked as runnable the code sample will need to start
with `d2.`. As the code examples are run within an already
initialized d2 and currently against the dhis2 dev preview server.

Also adds support for static pages to the docs. These are pages that
do not describe code directly and will be more explainatory pages.
These pages will be written in markdown, which allows us to also make
them readable on github or parse them for other means.

Closes #5