[system] Get a single system setting.
SystemSettings class added that allows you to retrieve single
system settings using the get(id) method. It tries to parse the
response using JSON.parse. If this fails the text value is returned.
Empty values will be treated as invalid as the server returns 200 OK
for values that do not exist as well.

References #13