[model] Add reference type for ref/collections
Adds a reference type to the modelValidation object for
collections of references. This can for one be used to
locate what type of objects are referenced and to load
the available options for the collection from the api.
The referenceType only gets added for collections that
have a "REFERENCE" as their itemPropertyType

References #19