[model] Keep track of dirty properties
The model now keeps track of which properties have been changed
by keeping a set of property names that set() has been called on.
This will allow the save method to build an object that consists
of only the updated properties so we can build the PATCH object.

References #8
References #12