Initial support for managing collections - WIP:
`ModelDefinitions.getModelDefinitions()` returns a singleton, which should
be used throughout D2 in stead of separate `ModelDefinitions` objects.
This breaks several tests at the moment.

Collection properties on D2 model instances are now instances of
`ModelCollectionProperty` in stead of plain arrays. These objects handle
keeping track of changes made to the collection, as well as saving these
changes back to the API.