[CurrentUser] canCreate functions
When asking for create permissions for certain object we now support
ignoring specific modelTypes. The defaultConfig now contains a list
of modelTypes that should always return false when asking for create
permissions for those modelTypes.
The more concrete usecase for this is that the API specifies that
categoryOptionCombos are allowed to be created by the user, however
these should never be created by the user directly as they are
generated based on category options.

References dhis2/maintenance-app#111