[Model] Properly handle attributeValues on save
attributeValues are marked as COLLECTION type properties
by the the schema, however they are treated differently from the
regular COLLECTION type values. For the attributeValues, because of
their dynamic nature we just copy the value as is, and therefore
do not need to run it through the special collection handling.