Use static DataStore.getDataStore method for init
Creating the singleton when evaluating the module would cause
problems. when trying to fake/mock the api for cases where fetch
might not be available or is mocked in the beforeEach of a test.
This also bring the retrieval of the dataStore in line with how it
works for e.g. SystemSettings.