Create the `this.modelDefinition.modelValidations` in the constructor for ModelBase
There can be instances where a ModelBase is used without a proper Model,
and in that case most of the prototype functions work correctly, except
for the ones which look up `modelDefinition` and beyond.

`toJSON` is one of these functions. The expected response should be an
empty object if the subclass doesn't have any members so creating this
basic structure in the constructor works around the heisenbug which
appears from time to time.